refactor: Location 엔티티의 float값을 double값으로 변경

+++ src/main/java/com/ktc/togetherPet/model/dto/missing/MissingPetDTO.java
@@ -12,8 +12,8 @@ public record MissingPetDTO(
     long birthMonth,
     String petBreed,
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Seoul") LocalDateTime lostTime,
-    float latitude,
-    float longitude,
+    double latitude,
+    double longitude,
     String description,
     boolean isNeutering
 ) {
